Separate GUI from Wordlist class
